Preserve Android throttle samples before SDL integer narrowing

+++ src/joystick_calibration.h
@@ -10,12 +10,12 @@ struct JoystickCalibration {
     int center[2] = {0, 0};
     int high[2] = {32767, 32767};
     int throttleAxis = -1;
-    int idle = 32767;
-    int full = -32768;
+    double idle = 32767;
+    double full = -32768;
 };
 
 int joy_correctAxis(const JoystickCalibration &calibration, int axis, int value);
-int joy_correctThrottle(const JoystickCalibration &calibration, int value);
+int joy_correctThrottle(const JoystickCalibration &calibration, double value);
 void joy_loadCalibration(const std::string &path, int throttleAxis, JoystickCalibration &calibration);
 void joy_calibrationScreen(SDL_Joystick *joystick, int throttleAxis, JoystickCalibration &calibration);
